+++ casanovo/denovo/model_runner.py
@@ -110,6 +110,7 @@ def _execute_existing(
         max_charge=config["max_charge"],
         precursor_mass_tol=config["precursor_mass_tol"],
         isotope_error_range=config["isotope_error_range"],
+        n_beams=config["n_beams"],
         n_log=config["n_log"],
         out_writer=out_writer,
     )
@@ -261,6 +262,7 @@ def train(
         max_charge=config["max_charge"],
         precursor_mass_tol=config["precursor_mass_tol"],
         isotope_error_range=config["isotope_error_range"],
+        n_beams=config["n_beams"],
         n_log=config["n_log"],
         tb_summarywriter=config["tb_summarywriter"],
         warmup_iters=config["warmup_iters"],